decrease code duplication
